libgsmhr: Make it safe to decode/encode different streams at once
Basically the reference code has a bunch of global state.
With some minimal patchihg (previous commit) we can ensure that all
that state is within .bss

So what we do it to save/restore the bss section between calls of the
reference code so we can process several in // and we don't have to
completely fix all reference to global state in the reference codec.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>