Add gapk testsuit
This testsuite takes a PCM auidio file and encodes from it every
supported format, and compares that output with a known sample (from an
earlier, trusted version of gakp, shipped as part of this project).

I then re-decodes this file to PCM and also compares that with a shipped
reference re-decode.