tests: add pq_rtp test
This test is intended to check the RTP source / sink operability.
To do this, two processing queues are being allocated:

  "generator": source/random -> sink/rtp
  "checker": source/rtp -> sink/checker

The first one generates some amount of random bytes (payload),
and stores them inside a buffer that is shared between both
queues.

After generation, a payload is being sent from the first
queue via an RTP sink, and then being received by the second
via an RTP source.

As both queues do use a shared buffer, the last item of the
second queue (named 'sink/checker') is able to compare a
received payload with expected.